backend/requirements_advanced.txt
# SmartPlanner v2.0 - Dependencias Avanzadas con IA
# =====================================================

# FastAPI y dependencias web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
alembic==1.13.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Utilidades de fecha y tiempo
python-dateutil==2.8.2

# Validación de email
email-validator==2.1.0

# Análisis de datos e IA
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Monitoreo del sistema
psutil==5.9.5

# Programación de tareas
schedule==1.2.0

# Logging avanzado
colorlog==6.7.0

# Requests para APIs
requests==2.31.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# Utilidades adicionales
click==8.1.7
rich==13.6.0
tqdm==4.66.1

# Configuración
configparser

# Validación de tipos
typing-extensions==4.8.0

# Utilidades de red
urllib3==2.0.7

# Manejo de archivos
aiofiles==23.2.1

# Cache
cachetools==5.3.2

# Utilidades de tiempo
pytz==2023.3

# Validación de esquemas
jsonschema==4.19.2

# Análisis de rendimiento
memory-profiler==0.61.0

# Manejo de archivos YAML
pyyaml==6.0.1

# Utilidades de regex
regex==2023.10.3

# Análisis estadístico
statsmodels==0.14.0

# Utilidades de configuración
dynaconf==3.2.4

# Utilidades de monitoreo
prometheus-client==0.19.0

# Utilidades de validación
validators==0.22.0

# Análisis de código estático
flake8==6.1.0

# Utilidades de testing avanzado
faker==20.1.0

# Análisis de cobertura
coverage==7.3.2

# Utilidades de parsing
pyparsing==3.1.1

# Utilidades de sistema de archivos
watchdog==3.0.0

# Análisis de memoria
pympler==0.9

# Utilidades de red
netaddr==0.9.0

# Utilidades de logging
structlog==23.2.0

# Utilidades de validación de datos
schema==0.7.5

# Utilidades de criptografía
cryptography==41.0.7

# Análisis de código
pycodestyle==2.11.1

# Utilidades de análisis de datos
tabulate==0.9.0

# Análisis de performance de base de datos
sqlparse==0.4.4

# Utilidades de validación de esquemas
marshmallow==3.20.1

# Análisis de código
mypy==1.7.1

# Utilidades de testing
mock==5.1.0

# Utilidades de análisis de logs
parse==1.19.1

# Análisis de archivos
chardet==5.2.0

# Base de Datos y ORM
psycopg2-binary==2.9.9

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Inteligencia Artificial y Machine Learning
joblib==1.3.2

# Monitoreo del Sistema
schedule==1.2.0
colorlog==6.8.0
structlog==23.2.0

# Análisis de Código
ast-decompiler==0.7.0
astunparse==1.6.3
autopep8==2.0.4

# Testing Avanzado
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
pytest-mock==3.12.0

# Optimización y Cache
redis==5.0.1
memory-profiler==0.61.0

# Utilidades de Desarrollo
rich==13.7.0
typer==0.9.0

# Logging y Monitoreo
loguru==0.7.2
prometheus-client==0.19.0
grafana-api==1.0.3

# Análisis de Datos
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Comunicación y APIs
aiohttp==3.9.1
websockets==12.0

# Procesamiento de Archivos
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
fpdf2==2.7.6

# Validación y Serialización
cerberus==1.3.5
jsonschema==4.20.0

# Utilidades de Fecha y Tiempo
pytz==2023.3
arrow==1.3.0

# Criptografía y Hashing
hashlib
secrets

# Procesamiento de Texto
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0

# Configuración y Variables de Entorno
toml==0.10.2
configparser

# Utilidades del Sistema
pathlib2==2.3.7
shutil
tempfile
glob2==0.7

# Networking y Conectividad
ping3==4.0.4
netifaces==0.11.0
speedtest-cli==2.1.3

# Compresión y Archivos
zipfile38==0.0.3
tarfile
gzip

# Procesamiento JSON y XML
ujson==5.8.0
xmltodict==0.13.0
lxml==4.9.3

# Utilidades de Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentación
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.4.8

# Análisis de Rendimiento
line-profiler==4.1.1
py-spy==0.3.14
scalene==1.5.26

# Utilidades de Concurrencia
asyncio
threading
multiprocessing
concurrent.futures

# Gestión de Procesos
supervisor==4.2.5
gunicorn==21.2.0

# Utilidades de Red
certifi==2023.11.17

# Análisis Estadístico
pingouin==0.5.3

# Visualización Avanzada
bokeh==3.3.2
altair==5.2.0

# Machine Learning Avanzado
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2

# Procesamiento de Imágenes
pillow==10.1.0
opencv-python==4.8.1.78

# APIs y Microservicios
flask==3.0.0
celery==5.3.4
kombu==5.3.4

# Bases de Datos NoSQL
pymongo==4.6.0
redis-py-cluster==2.1.3

# Utilidades de Email
smtplib

# Gestión de Archivos
send2trash==1.8.2

# Utilidades de Sistema Operativo
platform
os
sys
subprocess

# Análisis de Logs
logparser==0.8.2
grok==1.0.0

# Utilidades de Backup
duplicity==2.1.4
rsync

# Monitoreo de Recursos
resource
gc
tracemalloc

# Utilidades de Red Avanzadas
scapy==2.5.0
nmap==0.0.1

# Análisis de Seguridad
bandit==1.7.5
safety==2.3.5

# Utilidades de Debugging
pdb
ipdb==0.13.13
pudb==2023.1

# Gestión de Dependencias
pip-tools==7.3.0
pipdeptree==2.13.1

# Utilidades de Tiempo Real
asyncio-mqtt==0.16.1
websocket-client==1.6.4

# Análisis de Código Estático
pylint==3.0.3
vulture==2.10

# Utilidades de Configuración
omegaconf==2.3.0

# Procesamiento de Streams
kafka-python==2.0.2
pika==1.3.2

# Utilidades de Serialización
pickle5==0.0.12
dill==0.3.7

# Análisis de Memoria
tracemalloc
pympler==0.9

# Utilidades de Paralelización
ray==2.8.1
dask==2023.11.0

# Gestión de Configuración
python-decouple==3.8

# Utilidades de Validación
voluptuous==0.14.1
schema==0.7.5

# Análisis de Dependencias
pipreqs==0.4.13
pip-audit==2.6.1

# Utilidades de Desarrollo Web
jinja2==3.1.2
markupsafe==2.1.3

# Gestión de Sesiones
itsdangerous==2.1.2
flask-session==0.5.0

# Utilidades de Compresión
brotli==1.1.0
zstandard==0.22.0

# Análisis de Performance
py-heat==0.0.6
snakeviz==2.2.0

# Utilidades de Networking
dnspython==2.4.2
ipaddress

# Gestión de Archivos Temporales
tempfile
mktemp

# Utilidades de Hash y Checksum
hashlib
xxhash==3.4.1

# Análisis de Código Dinámico
dis
inspect
ast

# Utilidades de Sistema
signal
atexit
weakref

# Gestión de Memoria
gc
resource
mmap

# Utilidades de Concurrencia Avanzada
gevent==23.9.1
eventlet==0.33.3 